move the designer component to workbench layer (#17620)
