+++ src/sql/workbench/browser/designer/designerTabPanelView.ts
@@ -3,12 +3,12 @@
  *  Licensed under the Source EULA. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { DesignerTab } from 'sql/base/browser/ui/designer/interfaces';
+import { DesignerTab } from 'sql/workbench/browser/designer/interfaces';
 import { IPanelView } from 'sql/base/browser/ui/panel/panel';
 import { Table } from 'sql/base/browser/ui/table/table';
 import { Disposable } from 'vs/base/common/lifecycle';
 import * as DOM from 'vs/base/browser/dom';
-import { CreateComponentsFunc } from 'sql/base/browser/ui/designer/designer';
+import { CreateComponentsFunc } from 'sql/workbench/browser/designer/designer';
 
 const ButtonHeight = 30;
 const HorizontalPadding = 10;
