+++ src/sql/workbench/browser/designer/interfaces.ts
