+++ src/sql/workbench/contrib/tableDesigner/browser/tableDesignerEditor.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import 'vs/css!./media/tableDesignerEditor';
-import { Designer } from 'sql/base/browser/ui/designer/designer';
+import { Designer } from 'sql/workbench/browser/designer/designer';
 import { attachDesignerStyler } from 'sql/platform/theme/common/styler';
 import { TableDesignerInput } from 'sql/workbench/browser/editor/tableDesigner/tableDesignerInput';
 import *  as DOM from 'vs/base/browser/dom';
@@ -21,13 +21,11 @@ import { SaveTableChangesAction } from 'sql/workbench/contrib/tableDesigner/brow
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { IColorTheme, ICssStyleCollector, registerThemingParticipant } from 'vs/platform/theme/common/themeService';
 import { DesignerPaneSeparator } from 'sql/platform/theme/common/colorRegistry';
-import { TableDesignerTextEditor } from 'sql/workbench/contrib/tableDesigner/browser/tableDesignerTextEditor';
 
 export class TableDesignerEditor extends EditorPane {
 	public static readonly ID: string = 'workbench.editor.tableDesigner';
 
 	private _designer: Designer;
-	private _designerTextEditor: TableDesignerTextEditor;
 	private _saveChangesAction: SaveTableChangesAction;
 
 	constructor(
@@ -62,10 +60,7 @@ export class TableDesignerEditor extends EditorPane {
 		this._saveChangesAction.enabled = false;
 		actionbar.push(this._saveChangesAction, { icon: true, label: false });
 
-		this._designer = new Designer(designerContainer, (editorContainer) => {
-			this._designerTextEditor = this._instantiationService.createInstance(TableDesignerTextEditor, editorContainer);
-			return this._designerTextEditor;
-		}, this._contextViewService);
+		this._designer = new Designer(designerContainer, this._instantiationService, this._contextViewService);
 		this._register(attachDesignerStyler(this._designer, this.themeService));
 		this._register(registerThemingParticipant((theme: IColorTheme, collector: ICssStyleCollector) => {
 			const border = theme.getColor(DesignerPaneSeparator);
